% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_dosing.R
\name{emax_biomarker}
\alias{emax_biomarker}
\title{Emax biomarker model}
\usage{
emax_biomarker(conc, b_baseline = 100, emax = 30, ec50 = 5)
}
\arguments{
\item{conc}{Drug concentration.}

\item{b_baseline}{Baseline biomarker level.}

\item{emax}{Maximum effect.}

\item{ec50}{Concentration giving half-maximal effect.}
}
\description{
Emax biomarker model
}
\examples{
conc <- seq(0, 20, by = 2)
emax_biomarker(conc, b_baseline = 100, emax = 30, ec50 = 5)
}
