% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subject_summaries}
\alias{subject_summaries}
\title{Extract subject-level summary (EB intercept and slope)}
\usage{
subject_summaries(fit)
}
\arguments{
\item{fit}{A \code{pmrm_nlme} object.}
}
\value{
Data frame with \code{id}, \code{b0} and \code{b1}.
}
\description{
Extract subject-level summary (EB intercept and slope)
}
