% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competing_risks.R
\name{summary.pmrm_competing}
\alias{summary.pmrm_competing}
\title{Summarize a competing-risks fit}
\usage{
\method{summary}{pmrm_competing}(object, ...)
}
\arguments{
\item{object}{A \code{pmrm_competing} object.}

\item{...}{Unused.}
}
\description{
Summarize a competing-risks fit
}
