% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_survival.R
\name{summary.pmrm_joint}
\alias{summary.pmrm_joint}
\title{Summarize a joint model fit}
\usage{
\method{summary}{pmrm_joint}(object, ...)
}
\arguments{
\item{object}{A \code{pmrm_joint} object.}

\item{...}{Unused.}
}
\description{
Summarize a joint model fit
}
