% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed_effect.R
\name{print.pmrm_delayed}
\alias{print.pmrm_delayed}
\title{Print method for delayed-effect fits}
\usage{
\method{print}{pmrm_delayed}(x, ...)
}
\arguments{
\item{x}{A \code{pmrm_delayed} object.}

\item{...}{Passed to print.}
}
\description{
Print method for delayed-effect fits
}
